Move BIII search to package net.imagej.search.biii
