+++ src/main/java/net/imagej/search/biii/BIIIeuSearcher.java
@@ -27,7 +27,7 @@
  * #L%
  */
 
-package org.scijava.search.web;
+package net.imagej.search.biii;
 
 import java.io.BufferedReader;
 import java.io.IOException;
